[miniflare] Rename local R2 worker to public.worker
"public" is a more appropriate name for the worker than "local", since
Miniflare is inherently local.
